requirements.txt
astroid==2.1.0
Django==2.1.5
djangorestframework==3.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pkg-resources==0.0.0
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
pytz==2018.9
six==1.12.0
typed-ast==1.1.1
wrapt==1.10.11
